requirements.txt
gensim==3.8.3
nltk==3.3
numpy==1.15.2
pyltp==0.2.1
progressbar33==2.4
